Rename TMPL.java to SMARTBNNBNNSimilarity.java. This class implements the bnn.bnn TFxIDF variant. Add relevant hook to RetrievalApp.java
